mv build & bin.js
