+++ bin/build.js
@@ -1,136 +1,21 @@
-#!/usr/bin/env node
-
 var hyperstream = require('hyperstream')
 var fromString = require('from2-string')
-var pinoColada = require('pino-colada')
 var parallel = require('run-parallel')
 var explain = require('explain-error')
 var concat = require('concat-stream')
 var mapLimit = require('map-limit')
 var purify = require('purify-css')
-var resolve = require('resolve')
 var mkdirp = require('mkdirp')
-var subarg = require('subarg')
 var from = require('from2')
 var path = require('path')
-var pino = require('pino')
 var pump = require('pump')
 var zlib = require('zlib')
 var fs = require('fs')
 
-var htmlMinifyStream = require('./lib/html-minify-stream')
-var inspect = require('./bin/inspect')
-var start = require('./bin/start')
-var bankai = require('./')
+var htmlMinifyStream = require('../lib/html-minify-stream')
+var bankai = require('../')
 
-var pretty = pinoColada()
-pretty.pipe(process.stdout)
-
-var argv = subarg(process.argv.slice(2), {
-  string: [ 'open', 'port', 'assets' ],
-  boolean: [ 'watch', 'verbose', 'help', 'version', 'debug', 'electron' ],
-  default: {
-    address: 'localhost',
-    assets: 'assets',
-    debug: false,
-    open: false,
-    port: 8080
-  },
-  alias: {
-    address: 'A',
-    assets: 'a',
-    css: 'c',
-    debug: 'd',
-    electron: 'e',
-    help: 'h',
-    html: 'H',
-    js: 'j',
-    open: 'o',
-    port: 'p',
-    verbose: 'V',
-    version: 'v',
-    watch: 'w'
-  }
-})
-
-var usage = `
-  Usage:
-    $ bankai <command> [options]
-
-  Commands:
-    <default>                      Run 'bankai start'
-    start <filename>               Start a bankai server
-    build <filename> <directory>   Compile and export files to a directory
-    inspect <filename>             Visualize the dependency tree
-
-    Options:
-      -a, --assets=<directory>  Serve static assets [default: assets]
-      -A, --address=<ip>      Ip address to listen [default: localhost]
-      -c, --css=<subargs>     Pass subarguments to sheetify
-      -d, --debug             Include sourcemaps [default: false]
-      -e, --electron          Enable electron mode for the bundler
-      -h, --help              Print usage
-      -H, --html=<subargs>    Pass subarguments to create-html
-      -j, --js=<subargs>      Pass subarguments to browserify
-      -o, --open=<browser>    Open html in a browser [default: system default]
-      -p, --port=<n>          Bind bankai to a port [default: 8080]
-      -V, --verbose           Include debug messages
-      -w, --watch <bool>      Toggle watch mode
-
-  Examples:
-    $ bankai index.js -p 8080            # start bankai on port 8080
-    $ bankai index.js --open             # open html in the browser
-    $ bankai -c [ -u sheetify-cssnext ]  # use cssnext in sheetify
-    $ bankai -j [ -t brfs ]              # use brfs in browserify
-    $ bankai build index.js dist/        # compile and export to dist/
-
-  Examples:
-    bankai example.js --open=firefox-aurora -p 3000
-    bankai example.js --debug -w false
-`
-
-main(argv)
-
-function main (argv) {
-  if ((argv._[0] !== 'build') && (argv._[0] !== 'start') && argv._[0] !== 'inspect') {
-    argv._.unshift('start')
-  }
-
-  if (argv.h) {
-    console.log(usage)
-    return process.exit()
-  } else if (argv.v) {
-    console.log(require('./package.json').version)
-    return process.exit()
-  }
-
-  var cmd = argv._[0]
-  var _entry = argv._[1] || 'index.js'
-  var localEntry = './' + _entry.replace(/^.\//, '')
-  var entry = resolve.sync(localEntry, { basedir: process.cwd() })
-  var outputDir = argv._[2] || 'dist'
-
-  var logLevel = argv.verbose ? 'debug' : 'info'
-  argv.log = pino({ name: 'bankai', level: logLevel }, pretty)
-
-  if (cmd === 'start') {
-    start(entry, argv, handleError)
-  } else if (cmd === 'build') {
-    build(entry, outputDir, argv, handleError)
-  } else if (cmd === 'inspect') {
-    inspect(entry, argv, handleError)
-  } else {
-    argv.log.error(usage)
-    return process.exit(1)
-  }
-
-  function handleError (err) {
-    if (err) {
-      if (argv.verbose) throw err
-      else argv.log.error(err)
-    }
-  }
-}
+module.exports = build
 
 function build (entry, outputDir, argv, done) {
   var log = argv.log
@@ -146,8 +31,8 @@ function build (entry, outputDir, argv, done) {
   var buffers = {}
   var assets = bankai(entry, argv)
   var files = fs.existsSync(path.join(process.cwd(), 'index.html'))
-              ? ['bundle.js', 'bundle.css']
-              : ['index.html', 'bundle.js', 'bundle.css']
+    ? ['bundle.js', 'bundle.css']
+    : ['index.html', 'bundle.js', 'bundle.css']
 
   // Write to disk during watch mode
   if (argv.watch) {
