+++ package.json
@@ -4,11 +4,11 @@
   "description": "DIY asset server",
   "main": "index.js",
   "bin": {
-    "bankai": "bin.js"
+    "bankai": "bin/index.js"
   },
   "scripts": {
-    "start": "./bin.js example.js --open -p 8080 -A 0.0.0.0",
-    "build": "NODE_ENV=production ./bin.js build example.js",
+    "start": "./bin/index.js example.js --open -p 8080 -A 0.0.0.0",
+    "build": "NODE_ENV=production ./bin/index.js build example.js",
     "deps": "dependency-check . && dependency-check . --extra --no-dev",
     "test": "standard && npm run deps",
     "test:cov": "standard && npm run deps"
@@ -65,7 +65,7 @@
   "devDependencies": {
     "bel": "^4.5.0",
     "choo": "^5.0.4",
-    "dependency-check": "^2.6.0",
+    "dependency-check": "~2.6.0",
     "is-html": "^1.0.0",
     "istanbul": "^0.4.4",
     "openport": "0.0.4",
